refactor: 메인페이지 앨범 컴포넌트와 앨범 상세 페이지 컴포넌트 폴더 분리
