+++ client/src/entities/albumCard/index.ts
@@ -1,0 +1,1 @@
+export { AlbumCard } from './ui/AlbumCard';
