+++ client/src/entities/albumCard/model/types.ts
@@ -1,0 +1,11 @@
+/**
+ * 메인 페이지 앨범 카드
+ */
+
+export interface AlbumCardData {
+  albumId: string;
+  albumName: string;
+  artist: string;
+  albumTags?: string;
+  jacketUrl: string;
+}
