+++ client/src/entities/albumDetail/types.ts
