+++ client/src/entities/message/types.ts
@@ -1,3 +1,6 @@
+/**
+ * 채팅 메세지
+ */
 export interface MessageData {
   userName: string;
   message: string;
