+++ client/src/widgets/albumList/index.ts
@@ -1,0 +1,1 @@
+export { AlbumList } from './ui/AlbumList';
