+++ client/src/widgets/albums/index.ts
@@ -1,4 +1,3 @@
-export { AlbumList } from './ui/AlbumList';
 export { CommentList } from './ui/CommentList.tsx';
 export { Playlist } from './ui/Playlist';
 export { AlbumArtist } from './ui/AlbumArtist';
